Local execution improvements
