+++ Sources/SpeziLLMLocal/Configuration/LLMLocalParameters.swift
@@ -20,8 +20,9 @@ public struct LLMLocalParameters: Sendable {
     
     /// Defaults of possible LLMs parameter settings.
     public enum Defaults {
-        public static let defaultLlama2SystemPrompt: String = {
-            String(localized: LocalizedStringResource("SPEZI_LLM_LOCAL_LLAMA_SYSTEM_PROMPT", bundle: .atURL(from: .module)))
+        /// Default system prompt for local LLMs.
+        public static let defaultSystemPrompt: String = {
+            String(localized: LocalizedStringResource("SPEZI_LLM_LOCAL_SYSTEM_PROMPT", bundle: .atURL(from: .module)))
         }()
     }
     
@@ -129,7 +130,7 @@ public struct LLMLocalParameters: Sendable {
     /// Is passed to the underlying llama.cpp model in order to configure the LLM.
     ///
     /// - Parameters:
-    ///   - systemPrompt: The to-be-used system prompt of the LLM enabling fine-tuning of the LLMs behaviour. Defaults to the regular Llama2 system prompt.
+    ///   - systemPrompt: The to-be-used system prompt of the LLM enabling fine-tuning of the LLMs behaviour. Defaults to the regular default chat-based LLM system prompt.
     ///   - maxOutputLength: The maximum output length generated by the Spezi `LLM`, defaults to `512`.
     ///   - addBosToken: Indicates wether the BOS token is added by the Spezi `LLM`, defaults to `false`.
     ///   - gpuLayerCount: Number of layers to store in VRAM, defaults to `1`, meaning Apple's `Metal` framework is enabled.
@@ -141,7 +142,7 @@ public struct LLMLocalParameters: Sendable {
     ///   - useMmap: Indicates if mmap should be used., defaults to `true`.
     ///   - useMlock: Forces the system to keep model in RAM, defaults to `false`.
     public init(
-        systemPrompt: String = Defaults.defaultLlama2SystemPrompt,
+        systemPrompt: String = Defaults.defaultSystemPrompt,
         maxOutputLength: Int = 512,
         addBosToken: Bool = false,
         gpuLayerCount: Int32 = 1,
